[BBC-9048] ItemInfo.MDX
