+++ src/itemInfo/ItemInfo.story.mdx
@@ -1,0 +1,41 @@
+import { Meta, Story, Preview, Props } from '@storybook/addon-docs/blocks'
+
+import { ItemInfo } from './index'
+
+import { ClockIcon } from '../icon/clockIcon'
+
+<Meta title="Items|ItemInfo" />
+
+# ItemInfo
+
+<Preview>
+  <Story name="Default">
+    <ItemInfo
+      mainTitle="Main title"
+      mainInfo="Main info"
+      icon={<ClockIcon />}
+      tag={<li />}
+      aria-label="Custom aria label"
+    />
+  </Story>
+</Preview>
+
+## Specifications
+
+...
+
+## Usage
+
+```jsx
+import { ItemInfo } from '@blablacar/ui-library/build/itemInfo'
+
+<ItemInfo
+  mainTitle="Main title"
+  mainInfo="Main info"
+  icon={<ClockIcon />}
+  tag={<li />}
+  aria-label="Custom aria label"
+/>
+```
+
+<Props of={ItemInfo} />
