+++ src/itemInfo/ItemInfo.unit.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { shallow } from 'enzyme'
 
-import { ItemInfo } from '../itemInfo'
+import { ItemInfo } from './index'
 
 const defaultProps = {
   mainInfo: 'Main information',
